+++ client/src/components/BulmaExampleTable.vue
@@ -1,0 +1,67 @@
+<template>
+  <div v-if="examples">
+    <BulmaInput 
+      icon="filter"
+      placeholder="filter"
+      v-model="search"
+    />    
+    <ul style="columns:2">
+      <li v-for="e in filteredItems" :key="e.name">
+        <router-link :to="{path:$route.path, hash:'#'+e.name.replaceAll(' ','_').toLowerCase()}"><strong>{{ e.short||e.name }}</strong></router-link>
+      </li>
+    </ul>
+
+
+    <div v-for="f in filteredItems" class="mt-5" :key="f.name">
+      <h2 class="subtitle" :id="f.name.replaceAll(' ','_').toLowerCase()" v-html="highlight(f.name)">
+      </h2>
+      <p v-if="f.description">
+        <VueShowdown :markdown="highlight(f.description)" flavor="github" :options="{ghCodeBlocks:true}" />
+      </p>
+      <highlight-code v-if="f.code"
+        :lang="f.language || 'javascript'"
+        :code="f.code"
+      />
+    </div>
+  </div>
+</template>
+<script>
+  import Vue from 'vue'
+  import BulmaInput from './../components/BulmaInput.vue'
+  export default{
+    name:"BulmaExampleTable",
+    props:['examples'],
+    components:{BulmaInput},
+    computed:{
+      filteredItems(){
+        return this.examples?.filter(x => (x.name?.toLowerCase().includes(this.search.toLowerCase()) || 
+            x.description?.toLowerCase().includes(this.search.toLowerCase()) 
+        ))
+      },
+      hasOutput(){
+        return this.filteredItems?.filter(x => x.output)?.length>0
+      },
+
+    },
+    data(){
+      return{
+        search: ""
+      }
+    },
+    methods:{
+      escapeRegExp(string){
+          return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'); // $& means the whole matched string
+      },   
+      highlight(x){
+        if(this.search && x){
+          var esc = this.search.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&'); // eslint-disable-line
+          var reg = new RegExp(esc, 'ig');
+          return x.replace(reg, '<span style="background-color:yellow">$&</span>');
+        }
+        return x
+      },
+    }
+  }
+</script>
+<style scoped>
+</style>